+++ z-paging/components/custom-nomore/custom-nomore.vue
@@ -1,8 +1,8 @@
 <template>
 	<view class="nomore">
 		<!-- 这里的图片请换成自己项目的图片 -->
-		<image class="nomore-image" mode="aspectFit" src="@/static/no_more.png"></image>
-		<view class="nomore-text">已经到达宇宙尽头啦~</view>
+		<image class="nomore-image" mode="widthFix" src="@/static/no_more.png"></image>
+		<text class="nomore-text">已经到达宇宙尽头啦~</text>
 	</view>
 </template>
 
@@ -14,12 +14,15 @@
 			};
 		}
 	}
+	
 </script>
 
 <style scoped>
 	.nomore{
 		height: 150rpx;
+		/* #ifndef APP-NVUE */
 		display: flex;
+		/* #endif */
 		flex-direction: column;
 		align-items: center;
 		padding: 10px 0px;