+++ z-paging/pages/custom-demo/custom-demo.vue
@@ -1,8 +1,8 @@
-<!-- 自定义下拉刷新与上拉加载演示 -->
+<!-- 自定义下拉刷新与上拉加载演示(vue) -->
 <template>
 	<view class="content">
 		<!-- 非页面滚动时这里的fixed建议设置为true，则无需设置z-paging的高度及其父view的高度 -->
-		<z-paging ref="paging" :fixed="true" :auto-show-back-to-top="true" :refresher-threshold="80" @query="queryList"
+		<z-paging ref="paging" fixed :auto-show-back-to-top="true" :refresher-threshold="80" @query="queryList"
 			:list.sync="dataList">
 			<!-- 需要固定在顶部不滚动的view放在slot="top"的view中，如果需要跟着滚动，则不要设置slot="top" -->
 			<tabs-view slot="top" @change="tabChange" :items="['测试1','测试2','测试3','测试4']"></tabs-view>